Unify frontend data handling into one Router group
