+++ webscrambles/src/main/kotlin/org/worldcubeassociation/tnoodle/server/webscrambles/routing/WcifDataHandler.kt
@@ -4,29 +4,67 @@ import io.ktor.application.call
 import io.ktor.request.receive
 import io.ktor.response.respond
 import io.ktor.routing.Routing
+import io.ktor.routing.get
 import io.ktor.routing.post
 import io.ktor.routing.route
+import kotlinx.serialization.json.json
 import org.worldcubeassociation.tnoodle.server.RouteHandler
 import org.worldcubeassociation.tnoodle.server.webscrambles.Translate
+import org.worldcubeassociation.tnoodle.server.webscrambles.plugins.EventPlugins
+import org.worldcubeassociation.tnoodle.server.webscrambles.plugins.FormatPlugins
 import org.worldcubeassociation.tnoodle.server.webscrambles.wcif.model.Competition
+import org.worldcubeassociation.tnoodle.server.webscrambles.zip.folder.PrintingFolder
 import java.util.*
 
 object WcifDataHandler : RouteHandler {
     override fun install(router: Routing) {
         router.route("wcif/data") {
-            post("fmc-competitor-languages") {
-                val wcif = call.receive<Competition>()
+            get("events") {
+                val eventData = EventPlugins.values().map {
+                    json {
+                        "id" to it.key
+                        "name" to it.description
+                        "format_ids" to it.legalFormats.map(FormatPlugins::key)
+                        "can_change_time_limit" to (it !in EventPlugins.ONE_HOUR_EVENTS)
+                        "is_timed_event" to (it !in EventPlugins.ONE_HOUR_EVENTS)
+                        "is_fewest_moves" to (it == EventPlugins.THREE_FM)
+                        "is_multiple_blindfolded" to (it == EventPlugins.THREE_MULTI_BLD)
+                    }
+                }
 
-                val competitorCountries = wcif.persons
-                    .map { it.countryIso2 }
-                    .map { it.isoString }
-                    .distinct()
+                call.respond(eventData)
+            }
 
-                val countryLocales = competitorCountries.flatMap { getTranslatedCountryLanguage(it) }
-                    .toSet() + Translate.DEFAULT_LOCALE
+            get("formats") {
+                val formatData = FormatPlugins.WCA_FORMATS.mapValues {
+                    json {
+                        "name" to it.value.description
+                        "shortName" to it.value.tag
+                    }
+                }
 
-                val detectedLocales = countryLocales.distinct().map { it.toLanguageTag() }
-                call.respond(detectedLocales)
+                call.respond(formatData)
+            }
+
+            route("fmc/languages") {
+                post("competitors") {
+                    val wcif = call.receive<Competition>()
+
+                    val competitorCountries = wcif.persons
+                        .map { it.countryIso2 }
+                        .map { it.isoString }
+                        .distinct()
+
+                    val countryLocales = competitorCountries.flatMap { getTranslatedCountryLanguages(it) }
+                        .toSet() + Translate.DEFAULT_LOCALE
+
+                    val detectedLocales = countryLocales.distinct().map { it.toLanguageTag() }
+                    call.respond(detectedLocales)
+                }
+
+                get("available") {
+                    call.respond(AVAILABLE_LANGUAGE_TAGS)
+                }
             }
         }
     }
@@ -34,13 +72,13 @@ object WcifDataHandler : RouteHandler {
     private val COUNTRY_ISO2_LOCALES = Locale.getAvailableLocales()
         .groupBy { it.country }
 
-    private val TRANSLATED_LANGUAGE_ISO2 = Translate.TRANSLATED_LOCALES
-        .map { it.language }.toSet()
+    private val AVAILABLE_LANGUAGE_TAGS
+        get() = PrintingFolder.FMC_LOCALE_AVAILABLE_TAGS.toSortedSet()
 
-    private fun getTranslatedCountryLanguage(iso2: String): List<Locale> {
+    private fun getTranslatedCountryLanguages(iso2: String): List<Locale> {
         val candidateLocales = COUNTRY_ISO2_LOCALES[iso2] ?: return emptyList()
 
-        val translatedLocales = candidateLocales.filter { it.language in TRANSLATED_LANGUAGE_ISO2 }
+        val translatedLocales = candidateLocales.filter { it in Translate.TRANSLATED_LOCALES }
 
         val localLanguages = translatedLocales.filter { Locale(it.language, iso2) in Translate.TRANSLATED_LOCALES }
         val generalLanguages = translatedLocales.flatMap { Translate.TRANSLATED_LOCALES.filter { l -> l.language == it.language } }
